+++ SiamCross/SiamCross/Models/Sensors/Du/Measurement/DuMeasurementData.cs
@@ -9,13 +9,13 @@ namespace SiamCross.Models.Sensors.Du.Measurement
         public DuMeasurementSecondaryParameters SecondaryParameters {get;set;}
         public List<byte> Echogram { get; }
         public int FluidLevel { get; }
-        public int AnnularPressure { get; }
+        public float AnnularPressure { get; }
         public int NumberOfReflections { get; }
         public DateTime Date { get; }
 
         public DuMeasurementData(List<byte> echogram, 
                                  int fluidLevel,
-                                 int annualPressure,
+                                 float annualPressure,
                                  int numberOfReflections,
                                  DateTime date,
                                  DuMeasurementSecondaryParameters secondaryParameters)
